Began our program with GUI
The program now starts by trying to connect to the Zybo board, which will answer with the sentence "hello from Zybo" and this is how the program knows that it should continue and switch to the "main program" GUI, which right now is empty. I have successfully added some tabs where the third one has a nice graph which i was thinking we could show some timing info from doing some cryptographic algorithms on the ZYBO with and without hardware acceleration.